backend/requirements.txt
xlrd~=2.0
openpyxl~=3.1
django~=4.2
djangorestframework~=3.14
django-cors-headers~=4.2
python-magic~=0.4
gunicorn~=21.2
mysqlclient~=2.2
requests~=2.31
celery~=5.3
pydantic~=2.0
python-telegram-bot~=20.6
pydantic_extra_types~=2.2
phonenumbers~=8.13
email-validator~=2.1
jdatetime~=4.1
